+++ src/models/public-auth/Token.class.ts
@@ -1,7 +1,5 @@
 import { TokenData } from '@models/public-auth/dto/Token.dto'
 
-// type TokenType = 'CARD' | 'DIRECT_DEBIT'
-
 export class Token {
   readonly description: string
   readonly createdBy: string
